[SOFT-212] Make test logging cross platform
The STM32F0xx platform has the default size of |int| and |unsigned int|
as 2 bytes, while the x86 platform has them at 4 bytes. This causes
problems when matching formatting to a type with a specific size when
we're building for two different platforms.

+++ projects/solar/test/test_logger.c
@@ -69,7 +69,7 @@ void test_logger_everything_set(void) {
     data_store_set(data_point, set_value);
   }
 
-  LOG_DEBUG("All data points expected with numeric value %u:\n", set_value);
+  LOG_DEBUG("All data points expected with numeric value %u:\n", (unsigned int)set_value);
   logger_process_event(&s_data_ready_event);
 }
 
@@ -82,7 +82,7 @@ void test_logger_signed_values(void) {
     data_store_set(data_point, unsigned_value);
   }
 
-  LOG_DEBUG("Signed data points expected with %d, unsigned data points with %u:\n", signed_value,
-            unsigned_value);
+  LOG_DEBUG("Signed data points expected with %d, unsigned data points with %u:\n",
+            (int)signed_value, (unsigned int)unsigned_value);
   logger_process_event(&s_data_ready_event);
 }
